Updated City version
Updated version is in 
SIR QUBO reformulate to cities and links.ipynb

Converted 1-D finite element model, to a set of cities with links. The effect of the disease on each city affects each other city.  Next step is to use properties of this model to create a QUBO for DWave.