Objective Function added
Objective function added to the SIR city formulation. This ojective function is converted to a QUBO. It is balanced so that the appropriate cities are locked down (0) or kept open (1) to minimize the object function. As time goes by and people become immune,  DWave results will show cities should stay open.